feat: impl backend auth on publication range

+++ backend/internal/service/services.go
@@ -1,6 +1,8 @@
 package service
 
 import (
+	"errors"
+
 	"github.com/google/uuid"
 	"github.com/reibomaru/todo-app/backend/internal/model"
 	"gorm.io/gorm"
@@ -23,23 +25,51 @@ func (s Services) FindTaskStatusByCompanyID(companyID uuid.UUID) ([]*model.TaskS
 	return s.model.FindTaskStatusByCompanyID(s.DB, companyID)
 }
 
-func (s Services) FindTaskByID(companyID uuid.UUID, userID uuid.UUID, taskID uuid.UUID) (*model.Task, error) {
-	return s.model.FindTaskByID(s.DB, companyID, userID, taskID)
+func (s Services) FindTaskByID(userID uuid.UUID, taskID uuid.UUID) (*model.Task, error) {
+	return s.model.FindTaskByID(s.DB, userID, taskID)
 }
 
 func (s Services) FindMembersByCompanyID(companyID uuid.UUID) ([]*model.User, error) {
 	return s.model.FindUsersByCompanyID(s.DB, companyID)
 }
 
-func (s Services) CreateTask(newTask *model.CreateTaskPayload) error {
+func (s Services) CreateTask(userID uuid.UUID, newTask *model.CreateTaskPayload) error {
+	user, err := s.model.FindUserByID(s.DB, userID)
+	if err != nil {
+		return err
+	}
+	if user.Role == model.Viewer {
+		return errors.New("operation forbidden")
+	}
 	return s.model.CreateTask(s.DB, newTask)
 }
 
-func (s Services) UpdateTask(taskID uuid.UUID, payload *model.UpdateTaskPayload) error {
+func (s Services) UpdateTask(userID uuid.UUID, taskID uuid.UUID, payload *model.UpdateTaskPayload) error {
+	user, err := s.model.FindUserByID(s.DB, userID)
+	if err != nil {
+		return err
+	}
+	if user.Role == model.Viewer {
+		return errors.New("operation forbidden")
+	}
+	task, err := s.model.FindTaskByID(s.DB, userID, taskID)
+	if err != nil {
+		return err
+	}
+	if payload.PublicationRange != nil && task.AuthorID != userID {
+		return errors.New("operation forbidden")
+	}
 	return s.model.UpdateTask(s.DB, taskID, payload)
 }
 
-func (s Services) DeleteTask(taskID uuid.UUID) error {
+func (s Services) DeleteTask(userID uuid.UUID, taskID uuid.UUID) error {
+	user, err := s.model.FindUserByID(s.DB, userID)
+	if err != nil {
+		return err
+	}
+	if user.Role == model.Viewer {
+		return errors.New("operation forbidden")
+	}
 	return s.model.DeleteTask(s.DB, taskID)
 }
